feat(eslint-config): remove consistent-return for typescript
See the following SO for context why this is OK https://stackoverflow.com/a/67652059/1413689. Also tidied up the buildBaseTypescript file and added comments to make the decisions clearer. This isn't a breaking change as we are changing a global enablement to a conditional enablement so it shouldn't introduce any new eslint errors.