fix(scripts): github actions runner doesn’t correctly set systemDir
Fallback to a brute force detection of the systemDir so we can properly try to figure out where the closest node_modules is. Issue found in https://github.com/tablecheck/tablekit/runs/7554949913?check_suite_focus=true

Also added defaulting verbose to true if the debug logger is turned on in github actions.